requirements.txt
asgiref==3.8.1
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-after-response==0.2.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
ruff==0.11.2
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
whitenoise==6.9.0
